services/chatbot/requirements.txt
# An example requirements file. If your function has other dependencies,
# add them below
functions-framework==3.*
google-cloud-language==2.*
pytest==6.*
coverage==7.5.1
